requirements.txt
joblib==1.2.0
lightgbm==3.3.2
matplotlib==3.5.1
numpy==1.23.5
pandas==1.4.4
PyYAML==6.0
scikit-learn==1.1.3
tensorflow==2.13.1
xgboost==1.6.2
